requirements-docker.txt
# Web Framework - Python 3.13 compatible
Flask==3.1.1
Werkzeug==3.1.3

# Database ORM - Updated for Python 3.13
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41

# NOTE: pyodbc excluded for Docker - SQLite only

# Supabase Integration - Latest compatible version
supabase==2.16.0

# HTTP Requests
requests==2.32.3

# JSON Web Token handling
PyJWT==2.10.1

# Typing extensions
typing-extensions==4.14.1

# Additional utilities
python-dotenv==1.0.1
Flask-CORS==5.0.0

# Production WSGI Server
gunicorn==22.0.0       # For Docker/Linux

# Alternative SQL Server driver for production
pymssql==2.3.1         # Lightweight alternative to pyodbc

# Development/Testing
pytest==8.3.4